implement early db parts

+++ unitracker-chsu/src/model/auditoriums.rs
@@ -9,7 +9,7 @@ pub struct Auditorium {
     pub number: String,
     pub build_name: String,
     pub build_id: i64,
-    pub height: f64,
-    pub length: f64,
-    pub width: f64,
+    pub height: f32,
+    pub length: f32,
+    pub width: f32,
 }
